requirements.txt
asgiref
attrs
Django
django-cors-headers
django-database-url
django-extensions
django-mongoengine
django-rest-framework-mongoengine
djangorestframework
djangorestframework_simplejwt
dnspython
drf-spectacular
inflection
jsonschema
jsonschema-specifications
mongoengine
PyJWT
pymongo
PyYAML
dj_database_url
referencing
rpds-py
psycopg2
sqlparse
uritemplate